requirements.txt
# --- Core Frameworks ---
streamlit==1.39.0
python-dotenv==1.1.0
PyPDF2==3.0.1
typing-extensions==4.14.1
pydantic==2.11.7

# --- LangChain Core & Integrations ---
langchain==0.3.26
langchain-core==0.3.74
langchain-community==0.3.26
langchain-text-splitters==0.3.9
langchain-huggingface==0.3.1
langchain-groq==0.3.7
langchain-pinecone==0.2.8

# --- Embeddings & AI Libraries ---
torch==2.5.1
sentence-transformers==3.2.1
transformers==4.55.2
huggingface-hub==0.34.4

# --- Pinecone Dependencies ---
pinecone==7.3.0
grpcio==1.68.0
google-api-core==2.23.0

# --- Utilities ---
numpy==2.1.3
requests==2.32.3
tqdm==4.67.1
